feat: Free port (#15)
* feat: Free port

* fix: Rename APIPort

* fix: Rename APIEndpoint

+++ listener/listener.go
@@ -26,7 +26,7 @@ type Registry interface {
 }
 
 type Configuration interface {
-	WebPort() uint64
+	WebPort() int
 }
 
 func New(r Registry, c Configuration) *Listener {
